+++ google/cloud/bigtable/data/__init__.py
@@ -20,7 +20,7 @@ from google.cloud.bigtable import gapic_version as package_version
 
 from google.cloud.bigtable.data._async.client import BigtableDataClientAsync
 from google.cloud.bigtable.data._async.client import TableAsync
-from google.cloud.bigtable.data._async._read_rows import ReadRowsAsyncIterator
+# from google.cloud.bigtable.data._async._read_rows import ReadRowsAsyncIterator
 from google.cloud.bigtable.data._async.mutations_batcher import MutationsBatcherAsync
 
 from google.cloud.bigtable.data.read_rows_query import ReadRowsQuery
@@ -66,7 +66,7 @@ __all__ = (
     "DeleteAllFromRow",
     "Row",
     "Cell",
-    "ReadRowsAsyncIterator",
+    # "ReadRowsAsyncIterator",
     "IdleTimeout",
     "InvalidChunk",
     "FailedMutationEntryError",
