+++ google/cloud/bigtable/data/_async/__init__.py
@@ -14,13 +14,13 @@
 
 from google.cloud.bigtable.data._async.client import BigtableDataClientAsync
 from google.cloud.bigtable.data._async.client import TableAsync
-from google.cloud.bigtable.data._async._read_rows import ReadRowsAsyncIterator
+# from google.cloud.bigtable.data._async._read_rows import ReadRowsAsyncIterator
 from google.cloud.bigtable.data._async.mutations_batcher import MutationsBatcherAsync
 
 
 __all__ = [
     "BigtableDataClientAsync",
     "TableAsync",
-    "ReadRowsAsyncIterator",
+    # "ReadRowsAsyncIterator",
     "MutationsBatcherAsync",
 ]
